fix: PHPCS error in test file
